Region Checker / Basic Weapon Cleaner
Added the bare bones for the clean weapon function; still need to have it tie directly to other parts of the code before it will be implemented fully.

Added basic formating to a newly created file, as well as the bones to add more formating later if I wish.

Lots of additional safety checks, including cleaning up of input and having user confirmations of performing desturctive actions.